Word On Dashboard: implementation of navigation between tabs + summary, statisticy, and settings components
