+++ app/client/components/dashboard/LastWords.jsx
@@ -5,48 +5,23 @@ function LastWords() {
   return (
     <>
       <Text className='mb-2 text-lg font-semibold text-primary-dark'>
-        Last Words
+        Last Words Acquired
       </Text>
-      <Text className='mb-1 text-xs font-semibold uppercase text-slate-400 dark:text-slate-500'>
-        10 last Spoken
-      </Text>
-      <View className='flex flex-row items-start my-3'>
-        <Text className='mr-3 text-md font-bold flex items-center text-primary-dark'>
+      <View className='flex flex-row items-start'>
+        <View className='flex flex-row flex-wrap items-center w-56 mr-3 font-bold text-md text-primary-dark'>
           <MaterialCommunityIcons name='heart' size={24} color='#E9281F' />
-          Banana
+          <Text className='mr-3'> Cow</Text>
           <MaterialCommunityIcons name='heart' size={24} color='#E9281F' />
-          Teddy
+          <Text className='mr-3'> Tiger</Text>
           <MaterialCommunityIcons name='heart' size={24} color='#E9281F' />
-          Dog
-        </Text>
-        <Text className='mr-2 text-md font-bold text-primary-dark'>
+          <Text className='mr-3'> Dog </Text>
           <MaterialCommunityIcons name='heart' size={24} color='#E9281F' />
-          Milk
-        </Text>
-      </View>
-      <View className='flex flex-row items-start mb-3'>
-        <Text className='mr-2 text-md font-bold text-primary-dark'>
+          <Text className='mr-3'>Noodles</Text>
           <MaterialCommunityIcons name='heart' size={24} color='#E9281F' />
-          Cat
-        </Text>
-        <Text className='mr-2 text-md font-bold text-primary-dark'>
+          <Text className='mr-3'>Fast</Text>
           <MaterialCommunityIcons name='heart' size={24} color='#E9281F' />
-          Banana
-        </Text>
-      </View>
-      <View className='flex flex-row items-start mb-3'>
-        <Text className='mr-2  text-md font-bold text-primary-dark'>
-          <MaterialCommunityIcons name='heart' size={24} color='#E9281F' />
-          Banana
-        </Text>
-        <Text className='mr-2 text-md font-bold text-primary-dark'>
-          <MaterialCommunityIcons name='heart' size={24} color='#E9281F' />
-          Banana
-        </Text>
-        <Text className='mr-2 text-md font-bold text-primary-dark'>
-          <MaterialCommunityIcons name='heart' size={24} color='#E9281F' />
-          Banana
-        </Text>
+          <Text className=''> Bear </Text>
+        </View>
       </View>
     </>
   );
