+++ app/client/components/dashboard/Navbar.jsx
@@ -4,7 +4,7 @@ import { MaterialCommunityIcons } from '@expo/vector-icons';
 
 import { useState } from 'react';
 
-const Navbar = () => {
+const Navbar = ({ setActiveTab }) => {
   const { authUser } = AuthData();
 
   const logout = () => {
@@ -17,21 +17,36 @@ const Navbar = () => {
     setDropdownOpen(!isDropdownOpen);
   };
 
+  const handleSummary = () => {
+    setActiveTab('Summary');
+  };
+
+  const handleStatistics = () => {
+    setActiveTab('Statistics');
+  };
+
+  const handleSettings = () => {
+    setActiveTab('Settings');
+  };
+
   return (
     <>
       <View className='flex flex-row px-6 py-0  w-11/12 min-w-[85%] justify-end'>
         <View className='flex flex-row items-center justify-between px-6 py-0 '>
-          <Pressable className='mx-4 p-2 bg-[#b6e08ca3] shadow-sm border border-primary-green rounded-xl'>
+          <Pressable
+            className='mx-4 p-2 bg-[#b6e08ca3] shadow-sm border border-primary-green rounded-xl'
+            onPress={handleSummary}
+          >
             <Text className='font-bold text-center text-l text-primary-dark '>
               Summary
             </Text>
           </Pressable>
-          <Pressable className='mx-6'>
+          <Pressable className='mx-6' onPress={handleStatistics}>
             <Text className='font-bold text-center text-l text-primary-dark hover:text-primary-green'>
               Statistics
             </Text>
           </Pressable>
-          <Pressable className='mx-6'>
+          <Pressable className='mx-6' onPress={handleSettings}>
             <Text className='font-bold text-center text-l text-primary-dark hover:text-primary-green'>
               Settings
             </Text>
@@ -45,9 +60,9 @@ const Navbar = () => {
             >
               <View className={'flex flex-row items-center'}>
                 <Image
-                  source={authUser.profilePhoto}
+                  source={{ uri: `${authUser.profilePhoto}` }}
                   alt='user avatar'
-                  className='w-14 h-14 rounded-full shadow-md'
+                  className='rounded-full shadow-md w-14 h-14'
                   sx={{ br: '50%' }}
                 />
                 <MaterialCommunityIcons
@@ -57,7 +72,7 @@ const Navbar = () => {
                 />
               </View>
               {isDropdownOpen && (
-                <View className='absolute bg-lightgrey border border-gray-300 rounded shadow-md top-16 r-0'>
+                <View className='absolute border border-gray-300 rounded shadow-md bg-lightgrey top-16 r-0'>
                   <Pressable
                     onPress={logout}
                     className='px-4 py-2 hover:bg-slate-50'
