+++ app/client/components/dashboard/NewWords.jsx
@@ -10,9 +10,9 @@ function NewWords() {
         Since last week
       </Text>
       <View className='flex flex-row items-start'>
-        <Text className='mr-2 text-3xl font-bold text-primary-dark '>123</Text>
+        <Text className='mr-2 text-3xl font-bold text-primary-dark '>6</Text>
         <Text className='text-sm font-semibold text-white px-1.5 bg-[#398052] rounded-full'>
-          +49%
+          +100%
         </Text>
       </View>
     </>
