+++ app/client/components/dashboard/Settings.jsx
@@ -1,0 +1,5 @@
+const Settings = () => {
+  return <div>Settings</div>;
+};
+
+export default Settings;
