+++ app/client/components/dashboard/Statistics.jsx
@@ -1,0 +1,5 @@
+const Statistics = () => {
+  return <div>Statistics</div>;
+};
+
+export default Statistics;
